quicker exit for as_independent and unevaluated dep part
quicker exit in as_independent; evaluate=False

One may exit more quickly when the expr is not of the
type desired. So there should not be a need to use
the following in a calling routine:

```
if foo.is_Mul:
    a, b = foo.as_independent(x, as_Add=False)
```

Just use

```
a, b = foo.as_independent(x, as_Add=False)
```

Also, since one may send unevaluated expressions to
the routine, the routine should not do evaluation of
the expressions being returned (at least in the x-
dependent part). I'm not sure about the independent
part.

```
>>> Add(x, -x, evaluate=False).as_independent(x)
(0, x - x)
```