Recursively translate in experimental_lambdify
This fixes a test failure "TypeError: unsupported operand type(s) for /: 'Pi'
and 'interval'" from plot_implicit(y - cos(pi / x)). The failure can be traced
to the changing of the printing of Eq(x, y) from "x == y" to "Eq(x, y)". This
causes experimental_lambdify() to take a different code path, based on a
function call instead of an expression. Translations like sympy.pi -> math.pi
were only happening in the translate_str() function. They are now done in
translate_func(). The interval arithmetic for plot_implicit() only works for
floats, hence the error message.

Guards to disable float_wrap_evalf and complex_wrap_evalf are needed here
because we don't want to wrap internal expressions in a function with float()
or complex() (this breaks plotting of Integral(log((sin(x)**2 + 1)*sqrt(x**2 + 1)),
(x, 0, y)) in the tests, for instance, because the internal symbolic
expressions in x are wrapped in float() or complex(), but x is a symbolic
bound symbol that only becomes numeric when the integral is computed).