Use Sum instead of summation in plotting tests
According to the comments, this is clearly supposed to test a Sum. But
summation was computing the sum in closed form for at least one instance here
(it was testing plotting harmonic(y) instead of the Sum). This makes the test
a little slower, but I believe it is what was intended to be tested. In the
other instance the sum was not computed in closed form but could be in the
future (it's the zeta function).