For AugAssign it will be better to use the string form of the operator
That way you do not accidentally write AugAssign(x, AddOp, y) instead of
AugAssign(x, AddOp(), y) (which is what the doctest originally did).

It might also be a good idea to only expose the singleton objects, or give
they a different name from their classes, as is done with singletons in the
core.