De-singletonize classes in sympy.codegen.ast
Just because a class is atomic (no args) doesn't mean it needs to be a
singleton. == from Basic will work just fine, and so will isinstance checks. S
shouldn't be cluttered up with relatively little used classes that have little
benefit from the performance or memory gains of being singletonized.