Move restrictions in CodeBlock regarding Assignments to CodeBlock.cse
The restrictions that every argument be an assignment and that duplicate
assignments are disallowed are only needed because of cse. We want CodeBlock
to be useable in a more general setting, though.

Also add an error for AugmentedAssignments, since they don't work the same way
as Assignments (the variable must already be assigned, meaning duplicate
assignments must be allowed).

These are all NotImplementedErrors. They can work, but some additional things
need to be implemented.