Issue 2708: Don't split integrals that can't be computed
Issue #2708 was (partially) fixed in commit 6fe73a1f by evaluating
separately the terms of the antiderivative. That gave rise to another
issue #10445, which was fixed in #10515 by removing the separation.

This commit aims to combine the merits of the two PRs by treating
separately those terms of the antiderivative that contain unevaluated
integral factors and evaluating the rest together in one sum.

Add tests for issue 2708