10430: use domain to define symbols as real in finite check
This is the one place where the symbol assumption partially
control the output. If a symbol is declared as finite, real, or
complex, those definitions are used to determine if the solution
to a Mul is the union of the solutions of the factors.

This should probably be changed so that the conditions necessary for
a solution are explicitly returned with the solution and not modified
by assumptions on the symbol.