Scrap the rest of sympy.codegen.ast for now
I don't need any of this stuff just yet, and I'm unsure if the way that it was
done is the best way to do it, and am wary of guessing without a good
use-case.

Stuff removed:

Datatype stuff: I'm particularly worried about this, and I am trying to avoid
doing type stuff
for now (side note: it seems to me rather that the datatypes should be
classes, like Double(x)).

Variable and its subclasses: the whole point of that was to attach datatypes
to variables (and I also don't get the InOutAssignment stuff; that seems too
Fortran specific and that it ought to be abstracted in a different way or even
done automatically).

FunctionDef: functions seem useful, but I also didn't port the routine stuff
from symcc either. I don't know what the most useful set of language agnostic
abstractions for these things are. We'll probably need something eventually,
but for now CodeBlock is a decent way to combine code expressions.

Import: This only makes sense if we can codegen an entire file, which, without
the other stuff, isn't going to happen any time soon.

Declare: Only useful if we deal with datatypes.

Return: Not useful without FunctionDef. I'm also not sure if this is the right
abstraction.