10401: evaluate Eq/Ne when the lhs - rhs evaluates
The condition that the lhs and rhs be complex was added
without a test originally and it is not clear what it is
there for. The equality can still evaluate if one side
is real and the other complex. The condition that both
sides be Expr guards against objects thinking they are
not zero (like MatAdd -- see issue #10728.

The test that was corrected wouldn't fail as it was, but
the inequality now evaluates so the simpler expression
was used.