Make reversed Ranges keep their order
Fixes #10606.

The logic in Range._intersect for intersecting with intervals is particularly
nasty, because of the corner cases with infinite ranges (ranges with infinite
starts are handled differently). I believe the test cases test all the proper
combinations, but the logic should still probably be checked, and it likely
can be simplified.

This changes the behavior of Range.__iter__ to raise an exception for ranges
with infinite starts, since ranges always iterate in order. See the discussion
on the issue for more information about this. I felt that it was better to
always iterate in order, and catch possible bugs if the range is reversed,
than to iterate backwards in this case. It's worth noting, however, that this
makes it more annoying to determine the values in an infinite range beyond the
two that are printed by the pretty printer. But to fix this, I think we should
add better support for slicing ranges.

Also added a better docstring to Range.