A. Did some corrections to docstring of sylvester, and some changes/additions to the general docstring.
B. Added the function res(f, g, x) to compute the resultant of f, g
by evaluating the determinant of sylvester(f, g, x, 1).

Added this function because all cas's return 0 for resultant(f, 0),
which imho is wrong because f and 0 do _not_ have a common divisor.
The new function returns a statement that the resultant is _not_ defined.