Small optimization to MatMul._entry
Don't call Sum.doit(), even if expand is set to True, if we know the bounds
are symbolic. There's no way the summation can evaluate to anything in that
case, so it just wastes time in the summation algorithms. This can result in a
performance hit if many MatMul indexed terms in an expression.