Allow TensorProduct for any ImmutableMatrix
Previously, TensorProduct only called matrix_tensor_product for
instances of Matrix, but the method can be made to work perfectly fine
for an instance of MatrixBase.

Now, TensorProduct will return an explicit matrix if all of its
arguments are instances of MatrixBase, and matrix_tensor_product will
work correctly with instances of MatrixBase.