Hide and test for warnings from the plotting tests
There were two warnings in the plotting tests:

    sympy/plotting/tests/test_plot.py[3] /Users/aaronmeurer/Documents/Python/sympy/sympy/sympy/plotting/experimental_lambdify.py:230: UserWarning: The evaluation of the expression is problematic. We are trying a failback method that may still work. Please report this as a bug.
      warnings.warn('The evaluation of the expression is'
    ...                                                                                                                 [OK]
    sympy/plotting/tests/test_plot_implicit.py[2] ./Users/aaronmeurer/Documents/Python/sympy/sympy/sympy/plotting/plot_implicit.py:83: UserWarning: Adaptive meshing could not be applied to the expression. Using uniform meshing.
      warnings.warn("Adaptive meshing could not be applied to the"

The first warning comes from plotting Integral(log((sin(x)**2 + 1)*sqrt(x**2 +
1)), (x, 0, y)) on (y, 1, 5). The problem is that the integral can't be
directly lambdified (actually experimental_lambdified) to something that can
be evaluated. It has to use the fallback method that uses evalf. I'm unclear
why this necessarily represents a bug. For now we just suppress this warning.

The second warning comes from a test that his this comment

    #Test plots which cannot be rendered using the adaptive algorithm
    #TODO: catch the warning.
    plot_and_save(Eq(y, re(cos(x) + I*sin(x))), name=name)

So it is explicitly testing the warning case. This adds code to test the
warning and suppress it from the test output.