add more arg checking to imageset
Although a Python function could be accepted, if
there is branching within, it will not be encapsulated
by calling the expression with a variable as is done.
And although unvariate Expr could be converted
to an Expr, there are issues with whether a single
symbol as in imageset(x, set) means the function is
Lambda(x, x) -- S.IdentityFunction -- or Lambda(_, x),
mapping to a variable.