Ability to pass arbitrary inputs from init_printing to the printer
This code should allow arbitrary inputs to be passed from init_printing
to the printer of choice.

Example:

>>> from sympy import *
>>> init_printing(pretty_printing=False, use_latex=True, mode="inline")
>>> Rational(7, 2)
$\frac{7}{2}$

The init_printing in this case passes mode="inline" to the latex printer
because mode is not one of its keyword arguments and gets passed on
through **settings. This could possibly reduce the need for a lot of the
other keyword inputs for the init_printing function, however, it seemed
like the ipython interactive printing code used those and that clean up
would take more investigation.

Note that when not using Ipython there is not currently code that
supports choosing the latex printer from init_printing. I plan to make a
pull request and open an issue immediately following this one, however,
that should fix this.