enhance simplification for Integers._eval_imageset
in master:

>>> i=iter(imageset(Lambda(x,2*x-3/S(7)),S.Integers))
>>> next(i)
-3/7
>>> next(i)
11/7
>>> next(i)
-17/7
>>> imageset(Lambda(x, 2*x + 3/S(7)), S.Integers)
ImageSet(Lambda(x, 2*x + 3/7), Integers())
>>> i=iter(imageset(Lambda(x,2*x+3/S(7)),S.Integers))
>>> next(i)
3/7
>>> next(i)
17/7
>>> next(i)
-11/7

With the changes of this commit, both imagesets start with
two positives and then begin alternating sign.

In addition, the Wild's now have an exclusion so they will
not match the symbol of interest - they truly test whether
the expression is linear in the symbol. The added test with
a negative assumption on the symbol fails in master.