Allow passing of arguments to latex printer in IPython
The ability to pass settings to the latex printer in the IPython code of
init_printing() has been added. Now latex can be given additional
arguments when working in IPython such as fold_frac_powers.

Example

In [1]: from sympy import *

In [2]: init_printing(fold_frac_powers=True)

Also added is a portion of code that removes any attempt to user specify
the mode of latex (example: mode="inline"). This is because the calls to
latex in the IPython printer already specify modes. The warning that
'mode' has been removed from the settings dictionary is sent using the
debug call just like the other portions of the init_printing code.