mechanics: modified KanesMethod to accept empty load
I have made changes to the kanes_method in KanesMethod to accept
a load that is either None, an empty tuple or a list containing one empty
tuple. The new fr will be calculated with a dummy tuple of the inertial
frame of the system and the zero vector in that frame. The new fr will be
a Matrix of zero column as desired.