mechanics: updated modified KanesMethod
Improvement from previous version. Got rid of previously hacky if-checks.
Modified _f_list_parser to handle cases where the load is free. However,
the case where the forcing list contains an empty tuple is yet to be handled.
But the case is extremely rare and does seem to be taken consideration
previously. Also a small fix to the documentation of _f_list_parser.