Partial velocities no longer require frame reexpression.
Fixes #10865

When computing a partial velocity (i.e. partial derivative of a velocity vector
with respect to a generalized speed), it is known a priori that the generalized
speed does not appear in the direction cosine matrices that relate any of the
unit vectors in the vector expression, so the partials can be computed without
the need to re-express the velocity vector in the desired reference frame.

To allow this to function a flag has been added to the Vector.diff() method
that specifies whether or not the user knows if the variable in question is or
isn't present in the direction cosine matrices. Secondly, the partial_velocity
function utilizes this flag to ensure that the partials properly compute even
if the orientation of the reference frames associated with the velocity vector
is not specified.

This change will allow users to more cleanly solve problems where the location
and orientation of the body with respect to the Newtonian frame is ignorable,
for example a simple satellite.