Added a partial velocity method to Point.
This method provides a functionality for computing the partial linear velocity
of a Point. A partial velocity is always with respect to a velocity vector
defined in a frame and generalized speeds. The Point object is aware of the
velocity vector in a complete sense, but not the generalized speeds. The user
must pass in the generalized speeds. This method is only really relevant and
useful in when forming the equations of motion with Kane's method.