init_printing: compatibility between no_global and **settings
Before the changes, a no_global True flag would replace the printer that
is to be used with a lambda function that calls the printer and manually
inputs specific arguments. This would cause problems when trying to pass
additional arguments to the printer through **settings because the new
printer (lambda function) only accepted a single input, expr. To fix
this, no_global no longer results in a lambda function with manual input
arguments but rather the arguments are added to the settings dictionary.
This way the arguments will be passed to the printer in addition to the
user specified arguments.

Note: The previous code made a distinction on passing of arguments to
the pretty printer vs other printers and this new method maintains that
distinction.