mechanics: change the kanes_equation api to be ideal
Switched the argument order of kanes_equation so that
now the body list is the first argument and loads the second.
Raise a deprecation warning and switch the argument if the first
item of body list is a tuple. I believe no additional check is needed
because the rest of the code already does the check.
Also updated all the unit test and added ignore deprecation warning
to all of them.