Add __all__ to sympy/calculus/__init__.py
This makes "from sympy import euler" import the euler function from
sympy.functions, not the euler submodule from sympy.calculus. This has been a
regression since SymPy 0.7.6.

I don't know if this can/should be tested, or if there's a better way to
maintain __all__. However, it's worth noting that in the way it is currently
written pyflakes will give warnings for an inconsistent __all__.