Add LLVM JIT callbacks for PyCuba integration.
Also add vectorized callbacks for cubature and pycuba.
(The '_v' variants of each callback type: 'cubature_v' and 'cuba_v')

The LoopCreator class helps creating the blocks and branches for loops.

The vector size for cubature is size_t, but all the internal loop variables
use 32 bit ints.  This could cause problems if there are very large vector
lengths.

Also, ctypes.c_size_t converts to a 64 bit int.  This may not work correctly
on a 32-bit system (untested).

There is a potential issue with a proliferation of supported callbacks.
It might be good to create an interface that lets users specify the
layout and functionality for other callbacks.  The CodeSignature object
could be a starting point to define that interface.