Simple transcendental equation with negative bases
This is fix for #11096. Previously transcendental equation with only
positive bases were being evaluated.

To account for the problem solve(Eq((-2)**x,4)), function _tsolve can
not evaluate transcendental equation with negative bases by
re-evaluating it after making the base positive.

I am not quite sure if the fix is robust. This change is done only for
the sake of solving the equations of similar type.

test case is added