physics vector point docstring clarifications
While learning to use the physics module I found some inconsistencies in
the docstrings of Point's methods. There are a couple of places where
the variables are defined out of order compared to the order they are
input. Also one had an input of otherpoint but defines an input called
point.