Physics.vector.frame.Referenceframe documentation error
I found an odd bug while learning to use the physics module. In the
documentation for the orientnew method of of the ReferenceFrame class,
the previous author attempted to reduce the amount of writing that was
required by adding the documentation of orient onto the end of the
orientnew docstring using `orientnew.__doc__ += orient.__doc__`. This
however caused the sections of orientnew that had subheadings to be
overwritten with orient's documentation. This results in one of the
inputs being incorrect (an input is defined that is not used and an
input that is used is not defined) and got rid of an example of
orientnew. Given that the documentation for orient appears directly
above orientnew I have decided to just remove the addition of orient's
documentation from orient new.