Remove `u_decode` from compatiblity
This is almost entirely made possible by the support of u"" strings,
since this is primarily used in testing.  The only location we actually
use this and need it is in Python 2 when we may get an object of type
`str` that is actually non-ascii.  In that case, we need to decode the
bytes of the string to a unicode string, but in Python 3 we already have
the (unicode) string.  IMO, despite needing this compatibility, I think
it is better to keep it pinned to that use case rather than have it in
the compatibility file, since we should try to do strings correctly,
here we just cannot control what the user does.