Sequences: New basic combinatorial sequences
This module implements some special sequences that commonly appear in
combinatorial contexts. The functions return the initial segment of
given size for these sequences.

The need for such a module arises from the fact that sequences often
can be computed much more efficiently (for example via recurrences)
than by list comprehension.

A convincing example is the computation of the Euler numbers.
To compute the first 500 nonzero Euler numbers with 'euler_sequence'
is more than 1000 times faster than with list comprehension of 'euler'.
(Benchmark with SymPy 1.0 and Python 3.5.)

Moreover the typical use case often requires the initial segment of
of sequences. For example Bernoulli numbers rarely come alone: their
typical use is in the Taylor-MacLaurin expansion of functions which
needs B0, B2, B4,..., B2n up to some n.

Contrary to the 'bernoulli' and 'euler' functions the sequences in
this module do not depended on the mpmath module.

Currently the following sequences are implemented:
    * andre_sequence
    * bell_sequence
    * bernoulli_sequence
    * euler_sequence
    * tangent_sequence

Example:

To compute the first 250 nonzero Bernoulli numbers with 'bernoulli_sequence'
is about 70 times faster than with list comprehension (and needs no mpmath).

[bernoulli(n) for n in range(0, 500, 2)]
1743.085 ms

bernoulli_sequence(250)
25.568 ms

More examples, benchmarks and details can be found at:
http://luschny.de/math/sympy/Proposals.html