Create the tables in the benchmarking database
This script will create the tables desired for the database where each
table represents a different test. The code will ignore existing tables
and so best use would be to add your table to the tables created rather
than altering existing code. Best practice will be to name any
additional tables similar to the name of the script for which it will hold
the test data.