Fix latex(S.Naturals0)
Previously it returned \mathbb{N_0}. Now it returns \mathbb{N}_0. The 0 should
not be blackboard font. Also, for whatever reason, at least for me, LaTeX
renders \mathbb{N_0} with a subscript \nvdash instead of 0. This seems odd,
and I'm not sure if there is not some issue on my computer, but regardless,
the 0 should not be blackboard.