Added models to the SymPy repository
Originally this code was a part of the PyDy repository but it was
discussed that it could be added to the SymPy repository for use with
benchmarking and testing. Some minor changes had to be made to make sure
no external dependencies were added. The functions in the code used to
return a pydy.system.System object but have been changed to return a
KanesMethod object now. Also the input order for the arguments to
.kanes_equations(...) had to be switched due to a recent change in the
api.