Added to the orientnew docstring
Took the examples from orient and transfered them to orient new. This
required a change to the first input argument but was otherwise the
same. The examples have also been test run to make sure that they work
without error. Also made a small PEP8 change in line folding and removed
an unused variable from the examples in the orient docstring.