Made some significant changes in parabola.py and test_parabola.py
In parabola.py :   Added some explanations in the docstrings for the less known properties, used S() to return the sympy type, removed a format call, removed an unused import of random, added a condition to check if focus is a point of directrix and changed the input type of equation() function.

In test_parabola.py: Added tests for raised errors, used sympy rationals and Line, Point, Point2D, Parabola are imported directly from sympy.