Fix for issue #8193 (KeyError in dsolve)
This commit fixes the issue in discussed in the previous commit (i.e.
that supplying a `funcs` argument to classify_sysode will always trigger
a KeyError exception). The issue occurs because the dictionary used to
track ODE order is expected to be pre-populated with initial order
values but is only actually populated when the `funcs` argument is None,
the fix chosen here is to instead populate the dictionary on the fly
when computing the ODE order.