Migrated property attributes to a base class
I migrated the property attributes from the equations of motion classes
to a base class. This will ensure that the equations of motion should
have the same basic interface for accessing the EOM regardless of which
method is used and will begin to reduce code redundancy.

I also added a method to the base class to check if the equations of
motion have been calculated yet. This check appeared in both methods as
copy pasted lines of code and so this has now been centralized for
convienence.