init_session: improved detection of current IPython session
Previously, to detect if there was a running IPython session,
init_session relied on a function called get_ipython, injected into the
global namespace by IPython. However, this function is not in scope
during initialization, particularly, it is not in scope during the
execution of startup scripts. This had the unfortunate consequence of
creating a new mainloop on top of the previous one when init_session was
called in a startup script. This commit changes the detection to use
IPython.terminal.interactiveshell.TerminalInteractiveShell.initialized,
a public-API method that works in this edge case as expected. Hence, it
becomes possible to use init_session in startup scripts.