Fixed issue 9480: Matrix.rank() incorrect results
This is a fix for the rank of a particular matrix,

  m = Matrix([[-5 + 5*sqrt(2), -5], [-5*sqrt(2)/2 + 5, -5*sqrt(2)/2]]),

being computed incorrectly with simplify=False. Previously, m.rank()
returned a value of 2 when the correct output should have been 1.

Digging deeper revealed that the incorrect rank was the result
of an incorrect rref for m. calling m.rref() wrongly resulted in
the 2x2 identity matrix when the correct result is:

  /                           \
  | 1   -5/(-5 + 5 * sqrt(2)) |
  |                           |
  | 0   0                     |
  \                           /

To fix this issue, I replaced