Changed printing of logic expressions
Fix for #11435.

Changed the functions _print_And() _print_Or() to print `&` and `|`
symbols in the expressions instead of `And` and `Or`.

Previously, printing module used its superclass to print `Not`
expression. Hence, a new function _print_Not() is created to change the
printing of `Not`expressions.

Previous tests are modified to account for new changes.