Heavily revised joint.py code
I reworked the code so a lot of the redundant code is placed in
joint.__init__ instead of being repeated. I also made it so each joint
has 4 frames and 4 points (one set for each mass center and the joint
location for each body). The docstrings do not necessarily reflect the
changes yet.