Fix for #11534
In case of equations with radicals, the result is a FiniteSet consisted of
checked solutions of the 'result' calculated inside _solve_radical helper
function.

If the equation to solve does not satisfy the requirements needed in order
to remove the radicals (defined inside sympy.solvers.unrad function),
'cov' is empty and, therefore, the 'result' gets to be a Complement object.
In such a case, the Comlplement needs to be returned, instead of a
FiniteSet object.

For example:

eq1 = -y + x/sqrt(-x**2 + 1)
and
eq2 = -y**2 + x**2/(-x**2 +1)

should have the same solution, which is

sol = Complement(FiniteSet(-y/sqrt(y**2 + 1), y/sqrt(y**2 + 1)), FiniteSet(-1, 1))