Explicitly set the Sphinx theme to classic
Newer versions of Sphinx have changed the default theme to alabaster. Our docs
were still rendering mostly the same because we copied and modified the
classic theme css in the _static directory (to change the colors to green).
However, it was causing the top navigation bar to not appear, as alabaster
doesn't have it.

See http://stackoverflow.com/q/39262189/161801.