MatExpr.args_cnc doesn't report 1 as coeff
MatExpr was reporting the args_cnc as ([1], [...]) for a Mul expression
but Expr does not report the [1] -- it gives []. By correcting this,
some code is simplified in cse.

Also, since there is a doctest showing how a coefficient of a MatExpr
can have a non-commutative entity, the comment regarding that was
removed.