TensorProduct: Associativity of the tensor product
 In several applications of the tensor product, one mathematical object can be
decomposed as a tensor product of sub-parts (for example spinors on product
spaces).

However, previous to this commit, the following happened

In [1]: from sympy import symbols, srepr

In [2]: from sympy.physics.quantum import TensorProduct as TP

In [3]: A,B,C = symbols('A,B,C', commutative = False)

In [4]: TP(A,B,C)
Out[4]: AxBxC

In [5]: srepr(TP(A,B,C))
Out[5]: "TensorProduct(Symbol('A', commutative=False), Symbol('B',
	commutative=False), Symbol('C', commutative=False))"

In [9]: TP(TP(A,B),C)
Out[9]: AxBxC

In [10]: srepr(TP(TP(A,B),C))
Out[10]: "TensorProduct(TensorProduct(Symbol('A', commutative=False),
	 Symbol('B', commutative=False)), Symbol('C', commutative=False))"

The TP(TP(A,B),C) illustrates some object that has been decompused as TP(A,B),
but it does not produce the same output -- that is, the implementation of
TensorProduct is not associative. But the mathematical tensor product is.

The above also causes a problem during multiplication: TP(TP(A,B),C) *
TP(A,B,C) cannot be carried out because of the apparent different tensor
product decompositions.

 This commit instead gives

In [1]: from sympy import symbols

In [2]: from sympy.physics.quantum import TensorProduct as TP

In [3]: A,B,C = symbols('A,B,C', commutative = False)

In [4]: TP(A,B,C) == TP(TP(A,B),C)
Out[4]: True

In [5]: TP(A,B,C) == TP(A,TP(B,C))
Out[5]: True

That is, the TensorProduct is now associative.