matrices:Matrix**exponent stays unevaluated for non-integer exponent
If exponent can be sympified to a sympy integer then matrix**exponent
is evaluated by default.Otherwise , it is stored in a MatPow object
and can be evaluated by calling the doit() method.

Some changes have also been done to other test functions which
earlier evaluated Rational and symbolic exponents by default.

Another new feature has been added : doit() on a MatPow object
evaluates for symbolic exponents as well.

 matrices : Matrix ** exponent stays unevaluated for non-integer exponent

If exponent can be sympified to a sympy integer then matrix**exponent
is evaluated by default.Otherwise , it is stored in a MatPow object
and can be evaluated by calling the doit() method.