Improved distance() function
Fixes #11617.
function zip() has limitation of creating tuples equal to number which
is minimum of the lengths of the list given. Hence when passing 2D and
3D points, only 2 tuples were being created. This problem can be solved
by using map() function(in python 2) but unfortunately map() function
in Python3 has different functionality than map() in python2.

To solve the problem, new empty arrays(p1 and p2) are created which is
of maximum length of both the `self` and `p` and elements are inserted
into the array. This gives us two lists of same size which can be used
properly by zip() function without loss of data.