Add PULL_REQUEST_TEMPLATE.md
This includes a simple message telling people to cross reference issues with
"Fixes #NNNN". The message is placed in angle brackets so that if someone
leaves it in, it will be invisible. I tested GitHub, and for single-commit
pull requests it still auto-populates the pull request comment with the commit
message (below the template).

Fixes #10620.