printing: changed str printing of logic expressions to use operators
Changed str printing of logic expressions to use operators &, |, ~
instead of And, Or, Not. I opted to use paranthesis around all
non-Atomic expressions to achieve correctness and clarity. However,
it is worth noting that this approach sometimes uses extraneous
parentheses, as it does not account for precedence among the logical
operators.

Example:

print(And(Not(x), Or(y, z)))

previously: And(Not(x), Or(y, z))
now: (y | z) & (~x)