Add a place option to hyperexpand
``hyperexpand()`` tries to represent Meijer G-functions in terms
other special functions. It first looks for hypergeometric series
expansions at zero and infinity, using Slater's theorem, and then
attempts to find closed form representations for the series.

If the number p of a-parameters of the G-function is smaller than
the number q of b-parameters, then the possible series expansions
are centered at zero and converge everywhere in the complex plane.
On the other hand, if q is smaller than p, then the expansions are
centered at infinity and undefined at zero.

If p and q are equal, there are hypergeometric expansions both at zero
and at infinity. The G-function is branched at 1 or -1, and therefore
the expansions have restricted domains of convergence.

A typical example is

    ``meijerg(((0, 1), ()), ((1/2,), (0,)), x**2)``

which appears in the integration of ``1/sqrt(1 + x**2)**3``.
It has p = q = 2 and two representations,

    ``x/sqrt(1 + x**2)`` and ``1/sqrt(1 + 1/x**2)``.

The former comes from an expansion in positive powers of x that is
convergent for |x| < 1. It is branched at i and -i (corresponding to
the branching point ``x**2 = -1`` of the G-function), and has a
natural analytic continuation with branch cuts from i to ooi and
from -i to -ooi along the imaginary axis.

The latter expression comes from an expansion in powers of  ``1/x**2``
that is convergent for |x| > 1. It is also branched at i and -i, and
the branch cut of its analytic continuation extends from -i to i along
the imaginary axis. In particular, it is not well defined in the
neighborhood of zero.

When there are two representations, ``hyperexpand`` will choose the one
that has smaller number of operations. However, expansions at infinity
are usually wrong for integration even when they are slightly simpler
in form.

This PR adds a hint, ``place``, to guide ``hyperexpand`` in the choice
when there are two expansions. It is used by ``meijerint_indefinite``
to find a serviceable antiderivative if one exists.

Fixes #7337
Fixes #10211
Fixes #11806