Don't attempt to normalize _mpf_ input to Float
This change was added in commit a6a6d66d7c025691b7ae96265d8ae492f970fa90
without much explanation. However, it was causing Float to lose precision when
given an _mpf_ tuple or mpf (with mpmath.mp.dps == the Float prec argument >
15), see the added test.

This removes a test of a non-normalized mpf tuple (even mantissa). Such tuples
are invalid (see
http://mpmath.org/doc/current/technical.html#representation-of-numbers). I
don't see much benefit in allowing them. Perhaps they should be normalized in
Float (I have not attempted to do this), but it needs to be done in a way that
does not lose precision.