Correctly normalize mpf/_mpf_ input to Float
This is a better fix than fdf7d4706153b62d6f92802dfe9af1b570564519. It also
adds back in a test that was removed from that commit, although in a different
way. The normalization of the _mpf_ with even mantissa loses a binary digit of
precision, so it isn't equal to the other values, but we still test that it
does normalize, and that the Float precision is set correctly.

Also modify mpf_norm to handle non-mpz input when the mpmath gmpy backend is
used.

There is another case at the end of Float.__new__ that is attempting to do
some normalization on unknown input types using mpmath.mpf(num)._mpf_, but I
am unclear when that case would actually be hit, so I've added a comment and
left it alone for now.