Add a prec option to nsolve and change it to return Float and Matrix
Just adding a prec option to nsolve without changing it to return Floats
doesn't seem like a very good idea, since mpmath has a global precision,
meaning we either don't reset it (bad), or we do reset it and the returned mpf
prints with less precision than what was asked (although the full precision
would be saved in the mpf itself).

It makes sense that nsolve, which is a SymPy wrapper around mpmath.findroot,
should return SymPy types. However, this is technically a backwards
incompatible change, so I'm unsure if it should be made, and if so, if we
should attempt some kind of deprecation (I don't know how to deprecate a
return type). The best I can think of would be to only return a Float if the
prec option is used, but that would be inconsistent and confusing.

This depends on #11862 but I'm submitting it separately as it might be
controversial, and I don't want it to block the fixes there.