Use existing normalization in Float._new
This finally goes back close to what was there before, except now we use evalf
on the fully symbolic expression (-1)**sign*man*2**exp (previously it did not
use a symbolic exponent and lost precision). We also now always normalize when
receiving a full length tuple, not just when the mantissa is 1, falling back
to evalf only when we don't get a full tuple (i.e., it doesn't have the fourth
precision term).

This reverts a test that I added in the previous commit, because I did not
change the Float._prec to match the _mpf_[-1] any more (making such a change
resulted in many other test failures, so I've left it alone for now, with some
comments).