Improved Latex Printing of Piecewise Functions
Fixes #11275 .

Pretty printer always encloses the Piecewise function when the
Piecewise function is multiplied to another function.

To match the output of Latex printer with that of Pretty Printer, a
conditional statement is added in the function `_needs_mul_brackets()`
to always include brackets for piecewise function whenever is
multiplied to another function.

Tests are changed to accommodate for the changes.